Updated the Data Collection file
As a result of the requirement to use the CRISP-DM methodology, it was necessary to separate responsibilities. I updated the code by moving the pip install commands to the requirements.txt file to reduce clutter in the notebook. I also removed the data cleaning and preparation code so that the cleaned data can be handled in a separate notebook (data_cleaning.ipynb).