Updated file with data visualization codes.
I added visualizations and plots to enhance data understanding. These include:

A bar chart showing the number of images in each subset of powdery and healthy cherry leaves.
A plot illustrating the distribution of image dimensions (height and width).
An image grid displaying sample images from both healthy and powdery leaf classes.
A visualization of the average image for each class.
A plot of the absolute difference between the average images of the two classes.